requirements.txt
SQLAlchemy==1.0.4
fake-factory==0.5.1
gnureadline==6.3.3
ipython==3.1.0
psycopg2==2.6
sqlalchemy-citext==1.3-0
